+++ CHANGELOG.md
@@ -6,10 +6,22 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `fpe::ff1::{InvalidRadix, NumeralStringError}`
+
 ### Changed
 - MSRV is now 1.56.0.
 - Bumped dependencies to `cipher 0.4`, `cbc 0.1`.
   - `aes 0.8` is now the minimum compatible crate version.
+- `fpe::ff1`:
+  - `FF1::new` now returns `Result<_, InvalidRadix>`.
+  - `FF1::{encrypt, decrypt}` now return `Result<_, NumeralStringError>`.
+  - `FF1::{encrypt, decrypt}` now enforce minimum and maximum lengths for
+    numeral strings. The minimum length depends on the radix, while the maximum
+    length is `u32::MAX` numerals (which means `BinaryNumeralString` can process
+    data that is less than 16 MiB).
+  - Refactored `NumeralString` trait:
+    - Renamed `NumeralString::len` to `NumeralString::numeral_count`.
 
 ## [0.5.1] - 2021-10-27
 ### Fixed